requirements.txt
discord.py==1.3.4
boto3==1.14.30
tweepy